// web/lib/api/models/bookableSlotDTO.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Reservae API
 * API para gerenciamento de espaços, disponibilidades e reservas.
 * OpenAPI spec version: v1
 */

export interface BookableSlotDTO {
  id?: number;
  /** @nullable */
  availabilityRuleId?: number | null;
  spaceId?: number;
  startsAt?: string;
  endsAt?: string;
  /** @nullable */
  customPricePerSpot?: number | null;
  capacity?: number;
  isActive?: boolean;
  reservedQuantity?: number;
}